refactor: organize components folder
