+++ src/app/(buzzer)/buzzer-box/page.tsx
@@ -2,9 +2,12 @@
 
 import dynamic from "next/dynamic";
 
-const BuzzerBox = dynamic(() => import("../../../components/pages/BuzzerBox"), {
-  ssr: false,
-});
+const BuzzerBox = dynamic(
+  () => import("../../../components/pages/buzzer/BuzzerBox"),
+  {
+    ssr: false,
+  },
+);
 export default function Page() {
   return <BuzzerBox />;
 }
