+++ src/app/(main)/(user-buzzer)/scorekeeper/page.tsx
@@ -3,10 +3,10 @@
 import dynamic from "next/dynamic";
 
 const Scorekeeper = dynamic(
-  () => import("../../../../components/pages/Scorekeeper"),
+  () => import("../../../../components/pages/buzzer/Scorekeeper"),
   {
     ssr: false,
-  }
+  },
 );
 export default function Page() {
   return <Scorekeeper />;
