+++ src/app/(main)/management/upload/page.tsx
@@ -1,4 +1,4 @@
-import UploadSet from "@/components/pages/UploadSet";
+import UploadSet from "@/components/pages/management/UploadSet";
 import { prismaClient } from "@/src/utils/clients";
 
 export default async function Page() {
