+++ src/app/(main)/study/math/page.tsx
@@ -3,10 +3,10 @@
 import dynamic from "next/dynamic";
 
 const MathPage = dynamic(
-  () => import("../../../../components/pages/MathPage"),
+  () => import("../../../../components/pages/study/ComputationalMath"),
   {
     ssr: false,
-  }
+  },
 );
 
 export default function Page() {
