+++ src/app/(main)/study/quiz/page.tsx
@@ -1,7 +1,7 @@
-import QuizPage from "@/components/pages/QuizPage";
+import QuickQuiz from "@/components/pages/study/QuickQuiz";
 import { prismaClient } from "@/src/utils/clients";
 
 export default async function Page() {
   const categories = await prismaClient.category.findMany();
-  return <QuizPage categories={categories} />;
+  return <QuickQuiz categories={categories} />;
 }
