+++ src/components/layout/QuizFinished.tsx
