+++ src/components/pages/buzzer/BuzzerBox.tsx
