+++ src/components/upload/steps/Uploading.tsx
@@ -1,4 +1,4 @@
-import { StepComponentProps } from "@/components/pages/UploadSet";
+import { StepComponentProps } from "@/components/pages/management/UploadSet";
 import {
   Button,
   Card,
