+++ src/utils/quiz-session-type-extension.ts
@@ -1,5 +1,14 @@
 import { ArrayElement } from "@/src/utils/array-utils";
-import { Prisma, PrismaClient } from "@prisma/client";
+import {
+  AlphabetRound,
+  Category,
+  CategoryTeamGroup,
+  Prisma,
+  PrismaClient,
+  Question,
+  Round,
+  ThemeRound,
+} from "@prisma/client";
 
 export type QuizSessionWithQuestions = NonNullable<
   Prisma.Result<
@@ -33,3 +42,47 @@ export type QuizSessionWithQuestions = NonNullable<
 export type QuestionWithRoundData = NonNullable<
   ArrayElement<QuizSessionWithQuestions["questionsTrackers"]>["question"]
 >;
+
+export function addRoundData(
+  question: Question,
+  round?: Round & {
+    themeRound?: ThemeRound | null;
+    alphabetRound?: AlphabetRound | null;
+    categoryTeamGroup?: CategoryTeamGroup | null;
+  },
+  category?: Category,
+): QuestionWithRoundData {
+  const emptyRound: Round & {
+    themeRound: ThemeRound | null;
+    alphabetRound: AlphabetRound | null;
+    categoryTeamGroup: CategoryTeamGroup | null;
+  } = {
+    id: -1,
+    name: "Undefined Round",
+    difficulty: -1,
+    themeRound: null,
+    alphabetRound: null,
+    categoryTeamGroup: null,
+    createdOn: new Date(),
+    modifiedOn: new Date(),
+  };
+
+  const emptyCategory: Category = {
+    id: -1,
+    name: "Undefined Category",
+    createdOn: new Date(),
+    modifiedOn: new Date(),
+  };
+
+  return {
+    ...question,
+    round: {
+      ...emptyRound,
+      ...round,
+    },
+    category: {
+      ...emptyCategory,
+      ...category,
+    },
+  };
+}
